+++ Sources/Danger/DangerUtils.swift
@@ -1,3 +1,4 @@
+import DangerShellExecutor
 import Foundation
 
 /// Utility functions that make Dangerfiles easier to write